(js) React is the way. HTML attributes aren’t even case insensitive :rage:
Attribute names for HTML elements may be written with any mix of lowercase and uppercase letters that are a case-insensitive match for the names of the attributes given in the HTML elements section of this document; that is, attribute names are case-insensitive.

TECHNICALLY this isn’t needed. However @abulte is more pleased when things are tidy so I’ll cave.
(he didn’t even see that `{{script}}` doesn’t have proper spacing !)
Did I just write a 30-lines-long commit message for a 3 chars commit on a friday evening ?

Signed-off-by: Grégoire Mougeolle <gregoire@nextmap.io>
Signed-off-by: Grégoire Mougeolle <gregoire@nextmap.io>