Fix user deletion performances issues
When deleting a user, the slug is set to `deleted` and then a free suffix is looked for in `populate_slug`.
It is reproductible on a local env with 20K+ users deleted (as in production).
The performance issues comes from the query for each suffix increase, leading to huge performances issues when iterating until 20K+ deleted-XXX slugs.

This fix gets all existing suffixes at once and look for free space efficiently