Add integrity checks on non existent list field instead of empty (#3242)
Fixes [#1430](https://github.com/datagouv/data.gouv.fr/issues/1430)

## First approach (reverted)

The main issue seems to be because of a [bug in
mongoengine](https://github.com/MongoEngine/mongoengine/issues/267).

This first draft is a hackish fix: when a `ListField` is updated, if
it's changed to be the empty list (or if the result is an empty list, if
you dropped the last element for example), instead of being set to `[]`,
it's in fact unset (the field is removed from the document).

The added test showcases this.

This happens in
[BaseDocument._delta](https://github.com/MongoEngine/mongoengine/blob/master/mongoengine/base/document.py#L766-L767):
if the value isn't different from the default value (which is `[]` by
default on a `ListField`), then the field is unset (instead of set to be
empty).

The hackish fix updates the `datasets` field to be `[]` in the
`post_save` signal. This means that we might want to do the same thing
for `Dataset.resources`, and maybe for other `ListField`s just in case.

We might want to follow another route, maybe subclassing
`MongoEngine.BaseDocument` or `MongoEngine.Document` to overwrite the
`_delta` method, or the
[`MongoEngine.Document._get_update_doc`](https://github.com/MongoEngine/mongoengine/blob/master/mongoengine/document.py#L528-L541)
method to not unset a field if it's a `ListField`.

For the record, there's an existing [opened
PR](https://github.com/MongoEngine/mongoengine/pull/2517/files) on
mongoengine (that I didn't review, so no clue if it fixes it properly),
but it's been open for 3 years already, so not sure if or when it'll be
merged.

## Second approach (implemented here)

For now, instead of overriding mongoengine, we add a check on
`check-integrity` command to spot issues with missing ListFields.
Since `check-integrity` runs exhaustively on models and their fields, it
allows to monitor issues on any ListField.
We may want to update our strategy if there is some update at some point
on the PR mentioned above.

---------

Co-authored-by: maudetes <maudet.estelle@gmail.com>
Co-authored-by: Thibaud Dauce <ThibaudDauce@users.noreply.github.com>