ci: do not publish on tags (#3455)
With `setuptools_scm` (introduced in
https://github.com/opendatateam/udata/pull/3434), the Pypi CI from the
`master` branch publish with the tag on the current commit (before it
always published with a suffix `.devx`).

Since the `11.1.1` and `11.1.1.dev0` are the same we don't need to
publish from `master` and from the tag. And since publishing from
`master` do the correct thing, we removed the CI on tags.