refactor!: remove linkchecker (#3483)
The goal is to remove the entrypoint system. I don't think we use this
linkchecker anymore (it's disabled in our settings), hydra is doing this
job now? We maybe need to check the `ResourceMixin@ExtrasField` (I just
copy/paste but it could be nice in another PR to add hydra types here?)

---------

Co-authored-by: maudetes <maudet.estelle@gmail.com>