refactor: standardization of test clients, CLI runner, mocker… (#3525)
I encounter a few bugs with the way we are testing / authenticating in
API tests during the upgrade of Flask / Flask-Security. This PR goal is
to merge when we have multiple ways of doing the same things to better
understand how requests, auth, cli calls are made to the application. I
switch from fixture to parent functions in inheritance because it's way
easier to understand what the function is doing by ctrl-clic than
magical fixtures…

- [x] Remove `TestClient` (replaced by functions inside
`APITestCaseMixin`)
- [x] Remove `ApiClient` (replaced by the same functions inside
`APITestCaseMixin`)
- [x] Remove `client` and `api` fixtures (use functions inside
`APITestCaseMixin`)
- [x] Switch to `with app.test_request_context(): login_user(user)` for
activities tests requiring authentication since it's not really an API
test…
- [x] Use `mocker` instead of `monkeypatch` (there was only one usage of
`monkeypatch` in the tests)
- [x] Remove `TemplateRecorder` since there is no more templates
- [x] Remove unused `httpretty` fixture
- [x] Remove `SitemapClient` since the sitemap is in cdata