fix: validating dict in boolean harvest filter (#3542)
Fix
https://errors.data.gouv.fr/organizations/sentry/issues/273692/?alert_rule_id=10&alert_timestamp=1764115213721&alert_type=email&environment=demo.data.gouv.fr&notification_uuid=44880ccd-2a42-443d-8e76-2e6e5edefe74&project=12&referrer=alert_email

Can be tested with https://www.datasud.fr/fr/indexer/service/ckan/ 

`value = extra["value"]` can already be a `dict` in Ckan, so
`json.loads() is failing`. The change is the `boolean` function is not
useful I think with the second change in spatial Ckan management but
it's better anyway to check for TypeError for `int()` parsing failing.